// src/app/models/financial/journal-entry-transaction-filter.ts
export class JournalEntryTransactionFilter {
    idJournalEntryTransaction: number = -1;
    idCompany: number = -1;
    idFiscalYear: number = -1;
    idStatus: number = -1;
    documentNumber: string = "";
    idTypeDate: number = -1;
    initDate: string = "";
    endDate: string = "";
    pageNumber = 1;
    pagelogs = 10;
}